# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Parameter E-commerece API",
        "description": "Documentation for apis that can be used and their responses",
        "version": "1.0.0"
    },
    "paths": {
...
        "/api/v1/products/{product}/track-stock-updates": {
            "get": {
                "tags": [
                    "Product"
                ],
                "summary": "index",
                "operationId": "47e1e25dd8573290bcfbe34e01dacfcf",
                "parameters": [
                    {
                        "name": "product",
                        "in": "path",
                        "description": "Product ID"
                    }
                ],
                "requestBody": {
                    "required": false
                },
                "responses": {
                    "200": {
                        "description": "Success",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "message": {
                                            "type": "string",
                                            "example": "Product stock updates retrieved successfully"
                                        },
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            },
                                            "collectionFormat": "multi",
                                            "example": []
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}